// src/models/person.ts
interface Person {
    id: string
    apartmentId: string
    firstName: string
    secondName: string
    middleName?: string
    gender: number
    dateOfBirth?: string
    phoneNumber?: string
    email?: string
    date: string
    ownershipPercentage: number
    remark?: string
}

export default Person